+++ CHANGELOG.md
@@ -2,4 +2,14 @@
 
 ## dev
 
-  * [TAG_NAME] change comment
+## v0.0.2 (2015-03-16)
+
+  * [Normalize] Now attachments is the first comment;
+  * [Normalize] Who is signed in sub task also becomes owner of a story;
+  * [Normalize] Subtask description also becomes comment of a story;
+  * [Normalize] Todas as histórias têm etiqueta com nome do projeto;
+  * [Normalize] In some projects, the sections become labels;
+
+## v0.0.1 (2015-03-15)
+
+  * [Asana] Dump full workspace to json file;